🎨 REV-1821 Update backend parameters

+++ data/src/main/java/com/multimoney/data/util/catalog/Gender.kt
@@ -1,12 +1,13 @@
 package com.multimoney.data.util.catalog
 
-sealed class Gender(val id: Int, val gender: String) {
+sealed class Gender(val id: Long, val gender: String) {
     object Female : SignUpStep(1, "Femenino")
     object Male : SignUpStep(2, "Masculino")
 
     object Search {
         fun getGenderList() = listOf(Female.name, Male.name)
 
-        fun getGenderIdByName(name: String) = if (Female.name == name) Female.id else Male.id
+        fun getGenderIdByName(name: String) =
+            (if (Female.name == name) Female.id else Male.id).toLong()
     }
 }
